requirements.txt
opencv-python>=4.8.0
pytesseract>=0.3.10
pillow>=9.5.0
numpy>=1.24.0
python-dotenv>=1.0.0